Added initial structures and motivations for EKF SLAM
Added some of the initialisations required for EKF SLAM, and following through the my understanding and motivation behind each line of the algorithm I'm trying to implement - following EKF_SLAM_known_correspondences from Probabilistic Robotics.

First using the version with known correspondences as it seems a bit more accessible to my level of understanding, and appears as though it can be generalised without completely redoing it to the version without the known landmark correspondences.