Added dimension back for signature in measurements
Added the third dimension back to landmarks and measurements, making the selector matrix f_x now (3 x 3n) horizontally stacked with a (3 x 3) identity matrix, after understanding its role when following through the EKF_SLAM_known_correspondences derivation, where I'd previously omitted it to only focus on the x and y coordinates of landmarks.

I've initialised every instance of the signature to 0 for now though, as I'm not currently using it, but it's there for generality.