Fixed bug in observation update step
Fixed a bug in the observation update where I'd set the state estimate and uncertainty update steps to never execute on the first measurement, due to an off-by-one iterator initialisation.

Then when running this code with the corrected iterator, bugs presented related to matrix multiplication, hence the fix on the measurement vector's formation, and on the Q matrix to not contain zero elements so that the later matrix inversion successfully computes.