Reduced size of noise quantities
Reduced the size of noise quantities, as there is very low uncertainty about the current pose and control data, as the supervisor API is being used. Would have put zeros in to reflect certainty given by the supervisor API, but these cause calculation issues with matrices in the observation update step.