Normalised headings & fixed pose bug
Normalised the headings to be within -pi and +pi radians, so that the state estimate doesn't jump erratically on account of headings growing outside of that range.

Also fixed a bug in the get_pose() function where the axis of rotation changing sign to indicate direction of rotation hadn't been addressed, causing the incorrect sign to be applied to the heading after it had rotated beyond one full rotation.