Tuned Q distance uncertainties
Tuned the Q distance uncertainties to 0.015m in both x and y directions for landmarks, to attempt to inaccuracies in the landmark recognition.

The cause of these inaccuracies is the bounding box - by setting it to the reasonable (or generous, depending on your perspective) size of the same radius as the object itself, the reported location of the centre of the landmark becomes inaccurate.

Whilst this can of course be fixed by getting better data - ie. setting the bounding box to be ludicrously small, such that it's nearly back to full precision, tuning the Q parameters to inform the EKF-SLAM observation update step that there's more uncertainty in the landmark measurements, to improve performance is of course the more faithful solution - hence, it's the one implemented here.

0.015m has been chosen because it's half the radius of the landmark objects, and appears to improve performance notably compared to what it was with the much lower value of 0.000000001m