A* Algorithm Complex - Semi-complete
Not complete, can't narrow down the causes yet but need to submit a commit for the sake of having a checkpoint. The issue as it stands is that the mapping of occupied cells which are hardcoded are translating wrong and likewise the origin of the robot in the occupied grid causing the returning path to overlap on occupied cells.