Fixed implementation issue that caused landmark estimate stability issues
Fixed an issue in the implementation of the observation update step, where previously, the stability of landmark estimates was poor, and would erratically jump between timesteps. After identifying the issue only presented when multiple measurements were processed in a single timestep, I looked into whether it was a parameter tuning issue with either the Q matrix or the ALPHA_NOISE constant, or an implementation issue. As shown in the correction, the implementation was at fault, where the previous batch-style update to the state estimate and covariance meant that each measurement was processed without the benefit of the correction step applied based on the previous measurement. Hence, by moving the finishing update code into the outer-for-loop, the issue has been resolved.